Transplant Apiritif script generator into Nose executor
