+++ bzt/engine.py
@@ -1087,7 +1087,7 @@ class SubprocessedExecutor(ScenarioExecutor):
         self._stderr_file = None
 
     def _start_subprocess(self, cmdline):
-        prefix = self.execution["executor"]
+        prefix = self.execution.get("executor", None) or "executor"
         self._stdout_file = self.engine.create_artifact(prefix, ".out")
         std_out = open(self._stdout_file, "wt")
         self.opened_descriptors.append(std_out)
