--- bzt/modules/apiritif.py
@@ -1,319 +1,0 @@
-"""
-Copyright 2017 BlazeMeter Inc.
-
-Licensed under the Apache License, Version 2.0 (the "License");
-you may not use this file except in compliance with the License.
-You may obtain a copy of the License at
-
-   http://www.apache.org/licenses/LICENSE-2.0
-
-Unless required by applicable law or agreed to in writing, software
-distributed under the License is distributed on an "AS IS" BASIS,
-WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
-See the License for the specific language governing permissions and
-limitations under the License.
-"""
-import os
-import re
-import sys
-from collections import OrderedDict
-
-from bzt import ToolError, TaurusConfigError
-from bzt.engine import ScenarioExecutor, Scenario
-from bzt.modules.aggregator import ConsolidatingAggregator
-from bzt.modules.functional import FunctionalAggregator
-from bzt.modules.selenium import FuncSamplesReader, LoadSamplesReader, SeleniumWidget
-from bzt.requests_model import HTTPRequest
-from bzt.six import string_types, iteritems
-from bzt.utils import get_full_path, shutdown_process, PythonGenerator, dehumanize_time, ensure_is_dict
-
-
-class ApiritifExecutor(ScenarioExecutor):
-    def __init__(self):
-        super(ApiritifExecutor, self).__init__()
-        self.plugin_path = os.path.join(get_full_path(__file__, step_up=2), "resources", "nose_plugin.py")
-        self.process = None
-        self.stdout_path = None
-        self.stderr_path = None
-        self.stdout_file = None
-        self.stderr_file = None
-        self.script = None
-        self.report_path = None
-        self.generated_script = None
-
-    def prepare(self):
-        scenario = self.get_scenario()
-        if "requests" in scenario:
-            self.script = self._generate_script(scenario)
-            self.generated_script = True
-        elif "script" in scenario:
-            self.script = self.get_script_path()
-            self.generated_script = False
-        else:
-            raise TaurusConfigError("You must specify either 'requests' or 'script' for Apiritif")
-
-        self.stdout_path = self.engine.create_artifact("apiritif", ".out")
-        self.stderr_path = self.engine.create_artifact("apiritif", ".err")
-        self.report_path = self.engine.create_artifact("report", ".ldjson")
-
-        if self.engine.is_functional_mode():
-            self.reader = ApiritifResultsReader(self.report_path, self.engine, self.log, [])
-            if isinstance(self.engine.aggregator, FunctionalAggregator):
-                self.engine.aggregator.add_underling(self.reader)
-        else:
-            self.reader = LoadSamplesReader(self.report_path, self.log, [])
-            if isinstance(self.engine.aggregator, ConsolidatingAggregator):
-                self.engine.aggregator.add_underling(self.reader)
-
-    def startup(self):
-        load = self.get_load()
-        executable = self.settings.get("interpreter", sys.executable)
-        nose_command_line = [executable, self.plugin_path, '--report-file', self.report_path]
-
-        if load.iterations:
-            nose_command_line += ['-i', str(load.iterations)]
-
-        if load.hold:
-            nose_command_line += ['-d', str(load.hold)]
-
-        nose_command_line += [self.script]
-
-        self.stdout_file = open(self.stdout_path, "wt")
-        self.stderr_file = open(self.stderr_path, "wt")
-        self.process = self.execute(nose_command_line, stdout=self.stdout_file, stderr=self.stderr_file)
-
-    def check(self):
-        if self.widget:
-            self.widget.update()
-
-        ret_code = self.process.poll()
-        if ret_code is not None:
-            if ret_code != 0:
-                with open(self.stderr_path) as fds:
-                    std_err = fds.read()
-                msg = "Apiritif %s (%s) has failed with retcode %s \n %s"
-                raise ToolError(msg % (self.label, self.__class__.__name__, ret_code, std_err.strip()))
-            return True
-        return False
-
-    def shutdown(self):
-        shutdown_process(self.process, self.log)
-        if self.stdout_file:
-            self.stdout_file.close()
-        if self.stderr_file:
-            self.stderr_file.close()
-
-    def post_process(self):
-        pass
-
-    def has_results(self):
-        return bool(self.reader and self.reader.read_records)
-
-    def get_widget(self):
-        if not self.widget:
-            self.widget = SeleniumWidget(self.script, self.stdout_path)
-        return self.widget
-
-    def _generate_script(self, scenario):
-        test_file = self.engine.create_artifact("test_api", ".py")
-        test_gen = ApiritifScriptBuilder(scenario, self.log)
-        test_gen.build_source_code()
-        test_gen.save(test_file)
-        return test_file
-
-
-class ApiritifScriptBuilder(PythonGenerator):
-    IMPORTS = """\
-import time
-import unittest
-
-import apiritif
-
-"""
-
-    def __init__(self, scenario, parent_logger):
-        super(ApiritifScriptBuilder, self).__init__(scenario, parent_logger)
-
-    def gen_setup_method(self):
-        keepalive = self.scenario.get("keepalive", None)
-        if keepalive is None:
-            keepalive = True
-
-        exc = TaurusConfigError("Apiritif required 'default-address' to be set")
-        default_address = self.scenario.get("default-address", exc)
-        base_path = self.scenario.get("base-path", None)
-        auto_assert_ok = self.scenario.get("auto-assert-ok", True)
-
-        setup_method_def = self.gen_decorator_statement('classmethod')
-        setup_method_def.append(self.gen_method_definition("setUpClass", ["cls"]))
-        target_line = "cls.target = apiritif.http.target(%r)" % default_address
-        setup_method_def.append(self.gen_statement(target_line, indent=8))
-
-        if base_path:
-            setup_method_def.append(self.gen_statement("cls.target.base_path(%r)" % base_path, indent=8))
-        setup_method_def.append(self.gen_statement("cls.target.keep_alive(%r)" % keepalive, indent=8))
-        setup_method_def.append(self.gen_statement("cls.target.auto_assert_ok(%r)" % auto_assert_ok, indent=8))
-        setup_method_def.append(self.gen_new_line(indent=0))
-        return setup_method_def
-
-    def build_source_code(self):
-        self.log.debug("Generating Test Case test methods")
-        imports = self.add_imports()
-        self.root.append(imports)
-        test_class = self.gen_class_definition("TestRequests", ["unittest.TestCase"])
-        self.root.append(test_class)
-        test_class.append(self.gen_setup_method())
-
-        for index, req in enumerate(self.scenario.get_requests()):
-            if not isinstance(req, HTTPRequest):
-                msg = "Apiritif script generator doesn't support '%s' blocks, skipping"
-                self.log.warning(msg, req.NAME)
-                continue
-
-            mod_label = re.sub('[^0-9a-zA-Z]+', '_', req.url[:30])
-            method_name = 'test_%05d_%s' % (index, mod_label)
-            test_method = self.gen_test_method(method_name)
-
-            self._add_url_request(req, test_method)
-
-            test_class.append(test_method)
-            test_method.append(self.gen_new_line(indent=0))
-
-    def _add_url_request(self, request, test_method):
-        """
-        :type request: bzt.requests_model.HTTPRequest
-        """
-        named_args = OrderedDict()
-
-        method = request.method.lower()
-        think_time = dehumanize_time(request.priority_option('think-time', default=None))
-
-        named_args['timeout'] = dehumanize_time(request.priority_option('timeout', default='30s'))
-        named_args['allow_redirects'] = request.priority_option('follow-redirects', default=True)
-
-        headers = {}
-        scenario_headers = self.scenario.get("headers", None)
-        if scenario_headers:
-            headers.update(scenario_headers)
-        if request.headers:
-            headers.update(request.headers)
-        if headers:
-            named_args['headers'] = headers
-
-        merged_headers = dict([(key.lower(), value) for key, value in iteritems(headers)])
-        content_type = merged_headers.get('content-type', None)
-        if content_type == 'application/json' and isinstance(request.body, (dict, list)):  # json request body
-            named_args['json'] = {key: value for key, value in iteritems(request.body)}
-        elif method == "get" and isinstance(request.body, dict):  # request URL params (?a=b&c=d)
-            named_args['params'] = {key: value for key, value in iteritems(request.body)}
-        elif isinstance(request.body, dict):  # form data
-            named_args['data'] = list(iteritems(request.body))
-        elif isinstance(request.body, string_types):
-            named_args['data'] = request.body
-        elif request.body:
-            msg = "Cannot handle 'body' option of type %s: %s"
-            raise TaurusConfigError(msg % (type(request.body), request.body))
-
-        kwargs = ", ".join("%s=%r" % (name, value) for name, value in iteritems(named_args))
-
-        request_line = "response = self.target.{method}({url}, {kwargs})".format(
-            method=method,
-            url=repr(request.url),
-            kwargs=kwargs,
-        )
-        test_method.append(self.gen_statement(request_line))
-        self._add_assertions(request, test_method)
-        self._add_jsonpath_assertions(request, test_method)
-        self._add_xpath_assertions(request, test_method)
-        if think_time:
-            test_method.append(self.gen_statement('time.sleep(%s)' % think_time))
-
-    def _add_assertions(self, request, test_method):
-        assertions = request.config.get("assert", [])
-        for idx, assertion in enumerate(assertions):
-            assertion = ensure_is_dict(assertions, idx, "contains")
-            if not isinstance(assertion['contains'], list):
-                assertion['contains'] = [assertion['contains']]
-            subject = assertion.get("subject", Scenario.FIELD_BODY)
-            if subject in (Scenario.FIELD_BODY, Scenario.FIELD_HEADERS):
-                for member in assertion["contains"]:
-                    func_table = {
-                        (Scenario.FIELD_BODY, False, False): "assert_in_body",
-                        (Scenario.FIELD_BODY, False, True): "assert_not_in_body",
-                        (Scenario.FIELD_BODY, True, False): "assert_regex_in_body",
-                        (Scenario.FIELD_BODY, True, True): "assert_regex_not_in_body",
-                        (Scenario.FIELD_HEADERS, False, False): "assert_in_headers",
-                        (Scenario.FIELD_HEADERS, False, True): "assert_not_in_headers",
-                        (Scenario.FIELD_HEADERS, True, False): "assert_regex_in_headers",
-                        (Scenario.FIELD_HEADERS, True, True): "assert_regex_not_in_headers",
-                    }
-                    method = func_table[(subject, assertion.get('regexp', True), assertion.get('not', False))]
-                    line = "response.{method}({member})".format(method=method, member=repr(member))
-                    test_method.append(self.gen_statement(line))
-            elif subject == Scenario.FIELD_RESP_CODE:
-                for member in assertion["contains"]:
-                    method = "assert_status_code" if not assertion.get('not', False) else "assert_not_status_code"
-                    line = "response.{method}({member})".format(method=method, member=repr(member))
-                    test_method.append(self.gen_statement(line))
-
-    def _add_jsonpath_assertions(self, request, test_method):
-        jpath_assertions = request.config.get("assert-jsonpath", [])
-        for idx, assertion in enumerate(jpath_assertions):
-            assertion = ensure_is_dict(jpath_assertions, idx, "jsonpath")
-            exc = TaurusConfigError('JSON Path not found in assertion: %s' % assertion)
-            query = assertion.get('jsonpath', exc)
-            expected = assertion.get('expected-value', '') or None
-            method = "assert_not_jsonpath" if assertion.get('invert', False) else "assert_jsonpath"
-            line = "response.{method}({query}, expected_value={expected})".format(
-                method=method,
-                query=repr(query),
-                expected=repr(expected) if expected else None
-            )
-            test_method.append(self.gen_statement(line))
-
-    def _add_xpath_assertions(self, request, test_method):
-        jpath_assertions = request.config.get("assert-xpath", [])
-        for idx, assertion in enumerate(jpath_assertions):
-            assertion = ensure_is_dict(jpath_assertions, idx, "xpath")
-            exc = TaurusConfigError('XPath not found in assertion: %s' % assertion)
-            query = assertion.get('xpath', exc)
-            parser_type = 'html' if assertion.get('use-tolerant-parser', True) else 'xml'
-            validate = assertion.get('validate-xml', False)
-            method = "assert_not_xpath" if assertion.get('invert', False) else "assert_xpath"
-            line = "response.{method}({query}, parser_type={parser_type}, validate={validate})".format(
-                method=method,
-                query=repr(query),
-                validate=repr(validate),
-                parser_type=repr(parser_type),
-            )
-            test_method.append(self.gen_statement(line))
-
-    def gen_test_method(self, name):
-        self.log.debug("Generating test method %s", name)
-        test_method = self.gen_method_definition(name, ["self"])
-        return test_method
-
-
-class ApiritifResultsReader(FuncSamplesReader):
-    FIELDS_EXTRACTED_TO_ARTIFACTS = ["requestBody", "responseBody", "requestCookiesRaw"]
-
-    def __init__(self, filename, engine, parent_log, translation_table):
-        super(ApiritifResultsReader, self).__init__(filename, parent_log, translation_table)
-        self.engine = engine
-
-    def _write_sample_data_to_artifacts(self, sample_extras):
-        for file_field in self.FIELDS_EXTRACTED_TO_ARTIFACTS:
-            if file_field not in sample_extras:
-                continue
-            contents = sample_extras.pop(file_field)
-            if contents:
-                filename = "sample-%s" % file_field
-                artifact = self.engine.create_artifact(filename, ".bin")
-                with open(artifact, 'wb') as fds:
-                    fds.write(contents.encode('utf-8'))
-                sample_extras[file_field] = artifact
-
-    def read(self, last_pass=False):
-        for sample in super(ApiritifResultsReader, self).read(last_pass=last_pass):
-            self._write_sample_data_to_artifacts(sample.extras)
-            yield sample
