+++ bzt/modules/python.py
@@ -1,12 +1,17 @@
+import os
 import re
 import sys
+from collections import OrderedDict
 
-import os
 from bzt import ToolError, TaurusConfigError, TaurusInternalException
-
-from bzt.engine import SubprocessedExecutor, HavingInstallableTools
+from bzt.engine import SubprocessedExecutor, HavingInstallableTools, Scenario
+from bzt.modules.aggregator import ConsolidatingAggregator
+from bzt.modules.functional import FunctionalAggregator
+from bzt.modules.selenium import LoadSamplesReader, FuncSamplesReader
+from bzt.requests_model import HTTPRequest
 from bzt.six import parse, string_types, iteritems
-from bzt.utils import get_full_path, TclLibrary, RequiredTool, PythonGenerator, dehumanize_time, BetterDict
+from bzt.utils import get_full_path, TclLibrary, RequiredTool, PythonGenerator, dehumanize_time, BetterDict, \
+    ensure_is_dict
 
 
 class NoseTester(SubprocessedExecutor, HavingInstallableTools):
@@ -34,10 +39,14 @@ class NoseTester(SubprocessedExecutor, HavingInstallableTools):
 
     def __tests_from_requests(self):
         filename = self.engine.create_artifact("test_requests", ".py")
-        wdlog = self.engine.create_artifact('webdriver', '.log')
-        nose_test = SeleniumScriptBuilder(self.get_scenario(), self.log, wdlog)
-        self.generated_methods.merge(nose_test.build_source_code())
-        nose_test.save(filename)
+        test_mode = self.execution.get("test-mode", None) or "selenium"
+        if test_mode == "apiritif":
+            builder = ApiritifScriptBuilder(self.get_scenario(), self.log)
+        else:
+            wdlog = self.engine.create_artifact('webdriver', '.log')
+            builder = SeleniumScriptBuilder(self.get_scenario(), self.log, wdlog)
+        self.generated_methods.merge(builder.build_source_code())
+        builder.save(filename)
         return filename
 
     def install_required_tools(self):
@@ -54,7 +63,21 @@ class NoseTester(SubprocessedExecutor, HavingInstallableTools):
         run python tests
         """
         executable = self.settings.get("interpreter", sys.executable)
-        nose_command_line = [executable, self.plugin_path, '--report-file', self.execution.get("report-file")]
+
+        if "report-file" in self.execution:
+            report_file = self.execution.get("report-file")
+        else:
+            report_file = self.engine.create_artifact("report", ".ldjson")
+            if self.engine.is_functional_mode():
+                self.reader = FuncSamplesReader(report_file, self.engine, self.log, [])
+                if isinstance(self.engine.aggregator, FunctionalAggregator):
+                    self.engine.aggregator.add_underling(self.reader)
+            else:
+                self.reader = LoadSamplesReader(report_file, self.log, [])
+                if isinstance(self.engine.aggregator, ConsolidatingAggregator):
+                    self.engine.aggregator.add_underling(self.reader)
+
+        nose_command_line = [executable, self.plugin_path, '--report-file', report_file]
 
         load = self.get_load()
         if load.iterations:
@@ -65,6 +88,12 @@ class NoseTester(SubprocessedExecutor, HavingInstallableTools):
 
         nose_command_line += [self._script]
         self._start_subprocess(nose_command_line)
+
+    def has_results(self):
+        if isinstance(self.reader, LoadSamplesReader):
+            return bool(self.reader) and bool(self.reader.buffer)
+        elif isinstance(self.reader, FuncSamplesReader):
+            return bool(self.reader) and bool(self.reader.read_records)
 
 
 class TaurusNosePlugin(RequiredTool):
@@ -305,3 +334,186 @@ from selenium.webdriver.support.wait import WebDriverWait
             selector = selector[1:-1]
 
         return aby, atype, param, selector
+
+
+class ApiritifScriptBuilder(PythonGenerator):
+    IMPORTS = """\
+import time
+import unittest
+
+import apiritif
+
+"""
+
+    def __init__(self, scenario, parent_logger):
+        super(ApiritifScriptBuilder, self).__init__(scenario, parent_logger)
+
+    def gen_setup_method(self):
+        keepalive = self.scenario.get("keepalive", None)
+        if keepalive is None:
+            keepalive = True
+
+        exc = TaurusConfigError("Apiritif required 'default-address' to be set")
+        default_address = self.scenario.get("default-address", exc)
+        base_path = self.scenario.get("base-path", None)
+        auto_assert_ok = self.scenario.get("auto-assert-ok", True)
+
+        setup_method_def = self.gen_decorator_statement('classmethod')
+        setup_method_def.append(self.gen_method_definition("setUpClass", ["cls"]))
+        target_line = "cls.target = apiritif.http.target(%r)" % default_address
+        setup_method_def.append(self.gen_statement(target_line, indent=8))
+
+        if base_path:
+            setup_method_def.append(self.gen_statement("cls.target.base_path(%r)" % base_path, indent=8))
+        setup_method_def.append(self.gen_statement("cls.target.keep_alive(%r)" % keepalive, indent=8))
+        setup_method_def.append(self.gen_statement("cls.target.auto_assert_ok(%r)" % auto_assert_ok, indent=8))
+        setup_method_def.append(self.gen_new_line(indent=0))
+        return setup_method_def
+
+    def build_source_code(self):
+        methods = {}
+        self.log.debug("Generating Test Case test methods")
+        imports = self.add_imports()
+        self.root.append(imports)
+        test_class = self.gen_class_definition("TestRequests", ["unittest.TestCase"])
+        self.root.append(test_class)
+        test_class.append(self.gen_setup_method())
+
+        for index, req in enumerate(self.scenario.get_requests()):
+            if not isinstance(req, HTTPRequest):
+                msg = "Apiritif script generator doesn't support '%s' blocks, skipping"
+                self.log.warning(msg, req.NAME)
+                continue
+
+            if req.label:
+                label = req.label
+            elif req.url:
+                label = req.url
+            else:
+                raise TaurusConfigError("You must specify at least 'url' or 'label' for each requests item")
+
+            mod_label = re.sub('[^0-9a-zA-Z]+', '_', req.url[:30])
+            method_name = 'test_%05d_%s' % (index, mod_label)
+            test_method = self.gen_test_method(method_name)
+            methods[method_name] = label
+
+            self._add_url_request(req, test_method)
+
+            test_class.append(test_method)
+            test_method.append(self.gen_new_line(indent=0))
+
+        return methods
+
+    def _add_url_request(self, request, test_method):
+        """
+        :type request: bzt.requests_model.HTTPRequest
+        """
+        named_args = OrderedDict()
+
+        method = request.method.lower()
+        think_time = dehumanize_time(request.priority_option('think-time', default=None))
+
+        named_args['timeout'] = dehumanize_time(request.priority_option('timeout', default='30s'))
+        named_args['allow_redirects'] = request.priority_option('follow-redirects', default=True)
+
+        headers = {}
+        scenario_headers = self.scenario.get("headers", None)
+        if scenario_headers:
+            headers.update(scenario_headers)
+        if request.headers:
+            headers.update(request.headers)
+        if headers:
+            named_args['headers'] = headers
+
+        merged_headers = dict([(key.lower(), value) for key, value in iteritems(headers)])
+        content_type = merged_headers.get('content-type', None)
+        if content_type == 'application/json' and isinstance(request.body, (dict, list)):  # json request body
+            named_args['json'] = {key: value for key, value in iteritems(request.body)}
+        elif method == "get" and isinstance(request.body, dict):  # request URL params (?a=b&c=d)
+            named_args['params'] = {key: value for key, value in iteritems(request.body)}
+        elif isinstance(request.body, dict):  # form data
+            named_args['data'] = list(iteritems(request.body))
+        elif isinstance(request.body, string_types):
+            named_args['data'] = request.body
+        elif request.body:
+            msg = "Cannot handle 'body' option of type %s: %s"
+            raise TaurusConfigError(msg % (type(request.body), request.body))
+
+        kwargs = ", ".join("%s=%r" % (name, value) for name, value in iteritems(named_args))
+
+        request_line = "response = self.target.{method}({url}, {kwargs})".format(
+            method=method,
+            url=repr(request.url),
+            kwargs=kwargs,
+        )
+        test_method.append(self.gen_statement(request_line))
+        self._add_assertions(request, test_method)
+        self._add_jsonpath_assertions(request, test_method)
+        self._add_xpath_assertions(request, test_method)
+        if think_time:
+            test_method.append(self.gen_statement('time.sleep(%s)' % think_time))
+
+    def _add_assertions(self, request, test_method):
+        assertions = request.config.get("assert", [])
+        for idx, assertion in enumerate(assertions):
+            assertion = ensure_is_dict(assertions, idx, "contains")
+            if not isinstance(assertion['contains'], list):
+                assertion['contains'] = [assertion['contains']]
+            subject = assertion.get("subject", Scenario.FIELD_BODY)
+            if subject in (Scenario.FIELD_BODY, Scenario.FIELD_HEADERS):
+                for member in assertion["contains"]:
+                    func_table = {
+                        (Scenario.FIELD_BODY, False, False): "assert_in_body",
+                        (Scenario.FIELD_BODY, False, True): "assert_not_in_body",
+                        (Scenario.FIELD_BODY, True, False): "assert_regex_in_body",
+                        (Scenario.FIELD_BODY, True, True): "assert_regex_not_in_body",
+                        (Scenario.FIELD_HEADERS, False, False): "assert_in_headers",
+                        (Scenario.FIELD_HEADERS, False, True): "assert_not_in_headers",
+                        (Scenario.FIELD_HEADERS, True, False): "assert_regex_in_headers",
+                        (Scenario.FIELD_HEADERS, True, True): "assert_regex_not_in_headers",
+                    }
+                    method = func_table[(subject, assertion.get('regexp', True), assertion.get('not', False))]
+                    line = "response.{method}({member})".format(method=method, member=repr(member))
+                    test_method.append(self.gen_statement(line))
+            elif subject == Scenario.FIELD_RESP_CODE:
+                for member in assertion["contains"]:
+                    method = "assert_status_code" if not assertion.get('not', False) else "assert_not_status_code"
+                    line = "response.{method}({member})".format(method=method, member=repr(member))
+                    test_method.append(self.gen_statement(line))
+
+    def _add_jsonpath_assertions(self, request, test_method):
+        jpath_assertions = request.config.get("assert-jsonpath", [])
+        for idx, assertion in enumerate(jpath_assertions):
+            assertion = ensure_is_dict(jpath_assertions, idx, "jsonpath")
+            exc = TaurusConfigError('JSON Path not found in assertion: %s' % assertion)
+            query = assertion.get('jsonpath', exc)
+            expected = assertion.get('expected-value', '') or None
+            method = "assert_not_jsonpath" if assertion.get('invert', False) else "assert_jsonpath"
+            line = "response.{method}({query}, expected_value={expected})".format(
+                method=method,
+                query=repr(query),
+                expected=repr(expected) if expected else None
+            )
+            test_method.append(self.gen_statement(line))
+
+    def _add_xpath_assertions(self, request, test_method):
+        jpath_assertions = request.config.get("assert-xpath", [])
+        for idx, assertion in enumerate(jpath_assertions):
+            assertion = ensure_is_dict(jpath_assertions, idx, "xpath")
+            exc = TaurusConfigError('XPath not found in assertion: %s' % assertion)
+            query = assertion.get('xpath', exc)
+            parser_type = 'html' if assertion.get('use-tolerant-parser', True) else 'xml'
+            validate = assertion.get('validate-xml', False)
+            method = "assert_not_xpath" if assertion.get('invert', False) else "assert_xpath"
+            line = "response.{method}({query}, parser_type={parser_type}, validate={validate})".format(
+                method=method,
+                query=repr(query),
+                validate=repr(validate),
+                parser_type=repr(parser_type),
+            )
+            test_method.append(self.gen_statement(line))
+
+    def gen_test_method(self, name):
+        self.log.debug("Generating test method %s", name)
+        test_method = self.gen_method_definition(name, ["self"])
+        return test_method
