+++ tests/modules/test_SeleniumExecutor.py
@@ -11,7 +11,7 @@ from tests import BZTestCase, local_paths_config, __dir__
 
 from bzt.engine import ScenarioExecutor
 from bzt.modules.provisioning import Local
-from bzt.modules.selenium import NoseTester, AbstractSeleniumExecutor
+from bzt.modules.python import NoseTester
 from bzt.modules.selenium import SeleniumExecutor, LoadSamplesReader, LDJSONReader, FuncSamplesReader
 from bzt.six import StringIO
 from tests.mocks import EngineEmul
@@ -326,7 +326,7 @@ class TestReportReader(BZTestCase):
         self.assertEqual(len(items), 2)
 
     def test_func_reader(self):
-        reader = FuncSamplesReader(__dir__() + "/../selenium/report.ldjson", logging.getLogger(), None)
+        reader = FuncSamplesReader(__dir__() + "/../selenium/report.ldjson", EngineEmul(), logging.getLogger(), None)
         items = list(reader.read())
         self.assertEqual(4, len(items))
         self.assertEqual(items[0].test_case, 'testFailure')
